Handle direct I/O fallbacks for disk manager
