+++ src/storage/heap/mod.rs
@@ -1,0 +1,1 @@
+pub mod wal_codec;
