+++ src/storage/mod.rs
@@ -1,6 +1,7 @@
 pub mod codec;
 pub mod disk_manager;
 pub mod disk_scheduler;
+pub mod heap;
 pub mod heap_recovery;
 pub mod index;
 pub mod io;
